renaming dataframe.rs to data_ingest.rs. Settting up structure for writing file out
